chore(atoms): move Badge and Checkbox into Atoms folder
